Split mega test to different classes, handles infixLhs correctly
