+++ src/test/java/me/fponzi/tlaplusformatter/format/lexicon/InfixLHSTest.java
@@ -1,0 +1,7 @@
+package me.fponzi.tlaplusformatter.format.lexicon;
+
+import me.fponzi.tlaplusformatter.LexiconTest;
+
+public class InfixLHSTest extends LexiconTest {
+
+}
