+++ src/test/java/me/fponzi/tlaplusformatter/format/lexicon/TlaModuleTest.java
@@ -1,9 +1,9 @@
 package me.fponzi.tlaplusformatter.format.lexicon;
 
-import me.fponzi.tlaplusformatter.SanyTester;
+import me.fponzi.tlaplusformatter.LexiconTest;
 import org.junit.jupiter.api.Test;
 
-class TlaModuleTest extends SanyTester {
+class TlaModuleTest extends LexiconTest {
     @Test
     void testFormatModule() {
         var spec = "---- MODULE Spec ----\n======\n";
